0.4 - Question 4
Great question! I finished a simple answer, but then went down a rabbit hole of how to account for ties. I ended up reading lots of stackoverflow, worked with grok and tried separate examples using fake data to get it to work. Eventually I was able to streamline all the possibilities down to a concise nice answer for this use case.